requirements.txt
annotated-types==0.7.0
appdata==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colored==2.3.0
decorator==4.4.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagrapi==2.1.3
moviepy==1.0.3
numpy==2.2.3
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
to-requirements.txt==2.0.11
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2025.1.26
